Finished the resource layer

+++ src/main/java/com/theresourceroom/api/service/article/ArticleServiceWebImpl.java
@@ -3,6 +3,8 @@ package com.theresourceroom.api.service.article;
 import com.theresourceroom.api.dal.dao.ArticleDAO;
 import com.theresourceroom.api.models.Article;
 
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 public class ArticleServiceWebImpl implements ArticleService {
@@ -20,11 +22,17 @@ public class ArticleServiceWebImpl implements ArticleService {
 
     @Override
     public Article getArticleById(int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return articleDAO.getArticleById(id);
     }
 
     @Override
     public List<Article> getArticlesInTheme(int topic) {
+        if(topic < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return articleDAO.getArticlesInTheme(topic);
     }
 
@@ -35,11 +43,17 @@ public class ArticleServiceWebImpl implements ArticleService {
 
     @Override
     public int updateArticleWithId(String name, String link_story, String link_image, int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return articleDAO.updateArticleWithId(name, link_story, link_image, id);
     }
 
     @Override
     public int deleteArticleById(int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return articleDAO.deleteArticleById(id);
     }
 }
