+++ src/main/java/com/theresourceroom/api/service/tutorial/TutorialServiceWebImpl.java
@@ -3,6 +3,8 @@ package com.theresourceroom.api.service.tutorial;
 import com.theresourceroom.api.dal.dao.TutorialDAO;
 import com.theresourceroom.api.models.Tutorial;
 
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 public class TutorialServiceWebImpl implements TutorialService {
@@ -20,11 +22,17 @@ public class TutorialServiceWebImpl implements TutorialService {
 
     @Override
     public List<Tutorial> getAllTutorialsForTopic(int topic_id) {
+        if(topic_id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return tutorialDAO.getAllTutorialsForTopic(topic_id);
     }
 
     @Override
     public Tutorial getTutorialById(int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return tutorialDAO.getTutorialById(id);
     }
 
@@ -35,11 +43,17 @@ public class TutorialServiceWebImpl implements TutorialService {
 
     @Override
     public int deleteTutorial(int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return tutorialDAO.deleteTutorial(id);
     }
 
     @Override
     public int updateTutorial(String name, String link_image, String link_story, int id) {
+        if(id < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return tutorialDAO.updateTutorial(name, link_image, link_story, id);
     }
 }
